requirements.txt
# Botwave Dependencies
# Python 3.12+

# Telegram Bot (Required for plumbing_telegram_bot.py)
python-telegram-bot>=20.7

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Payments
stripe>=7.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0

# Async Support
asyncio-mqtt>=0.16.0
aioredis>=2.0.1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
isort>=5.12.0

# PDF Report Generation
fpdf2>=2.7.0
matplotlib>=3.8.0
numpy>=1.26.0
